// app/models/Medecin.ts
export interface Medecin {
    idMedecin: number;
    nomMedecin: string;
    cinMedecin: number;
    tel: string;
    prixConsultation: number;
    etatMedecin: number;
    libelleSpecialite: string;
    libelleService: string;
    nomDeletablissement: string;
    delegationMedecin:string;
}